refactor: replace request funcs to api folder
