+++ src/API/deleteAccount.ts
@@ -1,6 +1,6 @@
 import axios from "axios";
-import { URL_DELETE_ACCOUNT } from "./constants";
-import setRegister from "./storage";
+import { URL_DELETE_ACCOUNT } from "../utils/constants";
+import setRegister from "../utils/storage";
 
 export default function deleteAccount(token: string) {
   const config = {
