+++ src/API/logOut.ts
@@ -1,4 +1,4 @@
-import setRegister from "./storage";
+import setRegister from "../utils/storage";
 
 export default function logOut(name: string) {
   setRegister("", name, false);
