+++ src/API/sighUp.ts
@@ -1,7 +1,7 @@
 import axios, { AxiosError, AxiosResponse } from "axios";
 import { SignResponse } from "../types/types";
-import { URL_REGISTER } from "./constants";
-import setRegister from "./storage";
+import { URL_REGISTER } from "../utils/constants";
+import setRegister from "../utils/storage";
 
 export default async function signUp(
   name: string,
